test: Operations v2 component vitests
