+++ bciers/libs/shared/testConfig/src/mocks.ts
@@ -29,6 +29,7 @@ const usePathname = vi.fn();
 const useSearchParams = vi.fn();
 const useSession = vi.fn();
 const getServerSession = vi.fn();
+const fetchOperationsPageData = vi.fn();
 
 export {
   actionHandler,
@@ -39,4 +40,5 @@ export {
   usePathname,
   useSearchParams,
   useSession,
+  fetchOperationsPageData,
 };
